% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_data.R
\name{get_match_data}
\alias{get_match_data}
\title{get_match_data}
\usage{
get_match_data(x)
}
\arguments{
\item{x}{html_node}
}
\value{
tibble(date, patch, patch_link, win, blue_team, blue_team_link,  red_team, red_team_link)
}
\description{
Get basic match data
}
