% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_history.R
\name{get_match_details}
\alias{get_match_details}
\title{get_match_details}
\usage{
get_match_details(x)
}
\arguments{
\item{x}{html_node}
}
\value{
get_match_data, get_player_stats, get_match_stats, get_match_meta
}
\description{
Get meta data (error handled)
}
