% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_meta.R
\name{get_match_meta}
\alias{get_match_meta}
\title{get_match_meta}
\usage{
get_match_meta(x)
}
\arguments{
\item{x}{html_node}
}
\value{
tibble(match_url, api_url, game_hash, game_id, tournament_id, vod)
}
\description{
Get meta data (error handled)
}
