% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_stats.R
\name{get_match_stats}
\alias{get_match_stats}
\title{get_match_stats}
\usage{
get_match_stats(x)
}
\arguments{
\item{x}{html_node}
}
\value{
tibble(duration, blue_gold, blue_kills, blue_teamkills, blue_deths, blue_baron, blue_herald, red_gold, red_kills, red_teamkills, red_deaths, red_baron, red_herald, d_gold, d_kills, d_teamkills, d_deaths, d_baron, d_herald)
}
\description{
Get match stats (not error handled)
}
