% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_history.R
\name{get_matches}
\alias{get_matches}
\title{get_matches}
\usage{
get_matches(x)
}
\arguments{
\item{x}{html_node}
}
\value{
get_match_data, get_player_stats, get_match_stats, get_match_meta
}
\description{
Get all matches from a tournament
}
