% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_champion_acs.R
\name{get_player_champion_acs}
\alias{get_player_champion_acs}
\title{get_player_champion_acs}
\usage{
get_player_champion_acs(x)
}
\arguments{
\item{x}{html_node}
}
\value{
data
}
\description{
get players champion info
}
