% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_stats.R
\name{get_player_stats}
\alias{get_player_stats}
\title{get_player_stats}
\usage{
get_player_stats(x)
}
\arguments{
\item{x}{html_node}
}
\value{
players <- tibble(blue_players, blue_player_links, blue_champions, blue_bans, red_players, red_player_links,  red_champions, red_bans)
}
\description{
Get player stats data (basic error handler)
}
