% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournaments_pro.R
\name{get_tourn_match_list}
\alias{get_tourn_match_list}
\title{get_tourn_match_list}
\usage{
get_tourn_match_list(x)
}
\arguments{
\item{x}{a tournament row}
}
\value{
get_tournament_matches look for output
}
\description{
Get all matches from a tournament
}
