% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournaments_pro.R
\name{get_tourn_matches}
\alias{get_tourn_matches}
\title{get_tourn_matches}
\usage{
get_tourn_matches(data)
}
\arguments{
\item{data}{long tournament dataset}
}
\value{
get_tournament_matches look for output
}
\description{
Get all matches from a tournament
}
