% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournaments.R
\name{get_tournaments}
\alias{get_tournaments}
\title{get_tournaments}
\usage{
get_tournaments(x)
}
\arguments{
\item{x}{example url}
}
\value{
tibble(tourn_name, tourn_org, tourn_info, tourn_season, tourn_year, tourn_url)
}
\description{
Get tournaments avaible
}
