% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_data.R
\name{items_data}
\alias{items_data}
\title{Items information}
\usage{
items_data(api_key, region = "NA", data_locale = "en_US", version = NULL,
  itemListData = "all", tags = "all", dataById = "true")
}
\arguments{
\item{api_key}{API key provided by RIOT games (\url{https://developer.riotgames.com/})}

\item{region}{Summoner's region (RU, KR, BR, OCE, JP, NA, EUNE, EUW, TR, LAN, LAS)}

\item{data_locale}{Locale code for returned data (e.g., en_US, es_ES). If not specified, the default locale for the region is used.}

\item{version}{if not specified, it'll return the latest version information}

\item{itemListData}{Tags to return additional data. Only type, version, data, id, key, name, and title are returned by default if this parameter isn't specified. To return all additional data, use the tag 'all'.}

\item{tags}{Tags to return additional data. Only type, version, data, id, key, name, and title are returned by default if this parameter isn't specified. To return all additional data, use the tag 'all'.}

\item{dataById}{If specified as true, the returned data map will use the champions' IDs as the keys. If not specified or specified as false, the returned data map will use the champions' keys instead.}
}
\description{
This function allows you to get information about the items
}
\keyword{"API"}
\keyword{"Gaming"}
\keyword{"leagueOfLegends"}
