% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_data.R
\name{maps_data}
\alias{maps_data}
\title{Maps information}
\usage{
maps_data(api_key, region = "NA", data_locale = "en_US", version = NULL)
}
\arguments{
\item{api_key}{API key provided by RIOT games (\url{https://developer.riotgames.com/})}

\item{region}{Summoner's region (RU, KR, BR, OCE, JP, NA, EUNE, EUW, TR, LAN, LAS)}

\item{data_locale}{Locale code for returned data (e.g., en_US, es_ES). If not specified, the default locale for the region is used.}

\item{version}{if not specified, it'll return the latest version information}
}
\description{
This function allows you to get information about the maps
}
\keyword{"API"}
\keyword{"Gaming"}
\keyword{"leagueOfLegends"}
