% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_recent.R
\name{match_recent}
\alias{match_recent}
\title{Summoner's recent match information}
\usage{
match_recent(api_key, account_id, region = "NA")
}
\arguments{
\item{api_key}{API key provided by RIOT games (\url{https://developer.riotgames.com/})}

\item{account_id}{Summoner's account id}

\item{region}{Summoner's region (RU, KR, BR, OCE, JP, NA, EUNE, EUW, TR, LAN, LAS)}
}
\description{
This function allows you to get information about the last 20 matches for a player
}
\keyword{"API"}
\keyword{"Gaming"}
\keyword{"leagueOfLegends"}
