% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summoners_info.R
\name{summoners_info}
\alias{summoners_info}
\title{Summoner's basic information}
\usage{
summoners_info(api_key, summoners_name, region = "LAN",
  data_format = "json")
}
\arguments{
\item{api_key}{API key provided by RIOT games (https://developer.riotgames.com/)}

\item{summoners_name}{Summoner's name}

\item{region}{Summoner's region (RU, KR, BR, OCE, JP, NA, EUNE, EUW, TR, LAN, LAS)}

\item{data_format}{Data format on which the data will be output, available choices are json and data.frame}
}
\description{
This function allows you to get basic summoner's information.
}
\keyword{"API"}
\keyword{"league}
\keyword{legends"}
\keyword{of}
